+++ src/savefile.h
@@ -91,6 +91,7 @@ int rd_object_memory(void);
 int rd_quests(void);
 int rd_quests_1(void);
 int rd_quests_2(void);
+int rd_quests_3(void);
 int rd_artifacts(void);
 int rd_player(void);
 int rd_ignore(void);